requirements.txt
numpy==1.26.4
matplotlib==3.8.4
Pillow==10.3.0
PyWavelets==1.5.0
seaborn==0.13.2
ipykernel==6.29.4
scikit-image==0.22.0